Hash
This project is to create a hash table and use a data structure (I used
a templated linked list) where data is stored into the linked list. The
hash will hold they key (linked list id), the valie, and if there are
any collisions and stores it in the has table. The slot.h file is not
written by me. It was written by my professor: Dr. Raymer.