Blackjack Simulation
Simulation to answer two questions: What would the percentage difference be if the player was slightly riskier by stopping their hits at a score of 18 instead of 17 like the dealer? What percentage might the player have of winning if they followed to a strict strategy when playing against the dealer?